+++ src/utils/text.ts
@@ -2,4 +2,4 @@
  * Add thousands separators to a number.
  * @param number Number to format
  */
-export const formatNumber = (number: number) => number.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ' ');
+export const formatNumber = (number: number, separator = ' ') => number.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, separator);